+++ p2panda-rs/src/document/mod.rs
@@ -16,7 +16,9 @@
 //! # use p2panda_rs::hash::Hash;
 //! # use p2panda_rs::identity::KeyPair;
 //! # use p2panda_rs::operation::{OperationValue, OperationWithMeta};
+//! # use p2panda_rs::schema::SchemaId;
 //! # use p2panda_rs::test_utils::utils::{create_operation, delete_operation, update_operation, operation_fields};
+//! # use p2panda_rs::test_utils::constants::DEFAULT_SCHEMA_HASH;
 //! # use p2panda_rs::test_utils::mocks::{send_to_node, Client, Node};
 //! use p2panda_rs::document::DocumentBuilder;
 //! #
@@ -35,7 +37,7 @@
 //! #     .unwrap(),
 //! # );
 //! #
-//! # let schema = Hash::new_from_bytes(vec![3, 2, 1]).unwrap();
+//! # let schema = SchemaId::new(DEFAULT_SCHEMA_HASH).unwrap();
 //! # let mut node = Node::new();
 //! #
 //! # let (polar_entry_1_hash, _) = send_to_node(