+++ p2panda-rs/src/operation/operation.rs
@@ -9,6 +9,7 @@ use serde_repr::{Deserialize_repr, Serialize_repr};
 
 use crate::hash::Hash;
 use crate::operation::{OperationEncoded, OperationError, OperationFieldsError};
+use crate::schema::SchemaId;
 use crate::Validate;
 
 /// Field type representing references to other documents.
@@ -38,6 +39,11 @@ impl Relation {
                 false => Some(document_view),
             },
         }
+    }
+
+    /// Returns the relations document id
+    pub fn document_id(&self) -> &Hash {
+        &self.document
     }
 }
 
@@ -337,7 +343,7 @@ pub struct Operation {
     action: OperationAction,
 
     /// Hash of schema describing format of operation fields.
-    schema: Hash,
+    schema: SchemaId,
 
     /// Version schema of this operation.
     version: OperationVersion,
@@ -362,9 +368,9 @@ impl Operation {
     /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
     /// use p2panda_rs::hash::Hash;
     /// use p2panda_rs::operation::{AsOperation, Operation, OperationFields, OperationValue};
+    /// use p2panda_rs::schema::SchemaId;
     ///
-    /// let schema_hash_string = "0020c65567ae37efea293e34a9c7d13f8f2bf23dbdc3b5c7b9ab46293111c48fc78b";
-    /// let schema_msg_hash = Hash::new(schema_hash_string)?;
+    /// let msg_schema = SchemaId::new("0020c65567ae37efea293e34a9c7d13f8f2bf23dbdc3b5c7b9ab46293111c48fc78b")?;
     /// let mut msg_fields = OperationFields::new();
     ///
     /// msg_fields
@@ -374,14 +380,14 @@ impl Operation {
     ///     )
     ///     .unwrap();
     ///
-    /// let create_operation = Operation::new_create(schema_msg_hash, msg_fields)?;
+    /// let create_operation = Operation::new_create(msg_schema, msg_fields)?;
     ///
     /// assert_eq!(AsOperation::is_create(&create_operation), true);
     ///
     /// # Ok(())
     /// # }
     /// ```
-    pub fn new_create(schema: Hash, fields: OperationFields) -> Result<Self, OperationError> {
+    pub fn new_create(schema: SchemaId, fields: OperationFields) -> Result<Self, OperationError> {
         let operation = Self {
             action: OperationAction::Create,
             version: OperationVersion::Default,
@@ -397,7 +403,7 @@ impl Operation {
 
     /// Returns new UPDATE operation.
     pub fn new_update(
-        schema: Hash,
+        schema: SchemaId,
         previous_operations: Vec<Hash>,
         fields: OperationFields,
     ) -> Result<Self, OperationError> {
@@ -416,7 +422,7 @@ impl Operation {
 
     /// Returns new DELETE operation.
     pub fn new_delete(
-        schema: Hash,
+        schema: SchemaId,
         previous_operations: Vec<Hash>,
     ) -> Result<Self, OperationError> {
         let operation = Self {
@@ -447,7 +453,7 @@ pub trait AsOperation {
     fn action(&self) -> OperationAction;
 
     /// Returns schema of operation.
-    fn schema(&self) -> Hash;
+    fn schema(&self) -> SchemaId;
 
     /// Returns version of operation.
     fn version(&self) -> OperationVersion;
@@ -496,7 +502,7 @@ impl AsOperation for Operation {
     }
 
     /// Returns schema of operation.
-    fn schema(&self) -> Hash {
+    fn schema(&self) -> SchemaId {
         self.schema.to_owned()
     }
 
@@ -575,6 +581,7 @@ mod tests {
 
     use crate::hash::Hash;
     use crate::operation::OperationEncoded;
+    use crate::schema::SchemaId;
     use crate::test_utils::fixtures::templates::many_valid_operations;
     use crate::test_utils::fixtures::{fields, random_hash, schema};
     use crate::Validate;
@@ -628,7 +635,7 @@ mod tests {
     #[rstest]
     fn operation_validation(
         fields: OperationFields,
-        schema: Hash,
+        schema: SchemaId,
         #[from(random_hash)] prev_op_id: Hash,
     ) {
         let invalid_create_operation_1 = Operation {
@@ -700,7 +707,7 @@ mod tests {
 
     #[rstest]
     fn encode_and_decode(
-        schema: Hash,
+        schema: SchemaId,
         #[from(random_hash)] prev_op_id: Hash,
         #[from(random_hash)] document: Hash,
     ) {
@@ -741,7 +748,7 @@ mod tests {
     }
 
     #[rstest]
-    fn field_ordering(schema: Hash) {
+    fn field_ordering(schema: SchemaId) {
         // Create first test operation
         let mut fields = OperationFields::new();
         fields