+++ p2panda-rs/src/schema/mod.rs
@@ -13,11 +13,13 @@ use cddl::validator::cbor;
 mod cddl_builder;
 mod error;
 mod operation;
+mod schema_id;
 mod system_schema;
 
 pub use cddl_builder::CDDLBuilder;
 pub use error::{SchemaValidationError, SystemSchemaError};
 pub use operation::OPERATION_SCHEMA;
+pub use schema_id::SchemaId;
 
 /// Checks CBOR bytes against CDDL schemas.
 ///