+++ p2panda-rs/src/test_utils/fixtures/tests.rs
@@ -19,8 +19,9 @@ use crate::test_utils::fixtures::templates::{
     many_valid_entries, non_default_operation_values_panic, version_fixtures,
 };
 // Import dependencies for the templates module
+use crate::schema::SchemaId;
 use crate::test_utils::constants::DEFAULT_SCHEMA_HASH;
-use crate::test_utils::utils::{hash, operation_fields};
+use crate::test_utils::utils::operation_fields;
 
 // In this test `entry` and `key_pair` are injected directly into the test, they were imported from
 // the fixtures module above.
@@ -37,7 +38,7 @@ fn encode_entry(entry: Entry, key_pair: KeyPair) {
 #[case::default_operation(defaults::create_operation())]
 // This case should panic as we are passing in a non-default operation value.
 #[should_panic] // panic macro flag
-#[case::non_default_operation(create_operation(hash(DEFAULT_SCHEMA_HASH), operation_fields(vec![("message", OperationValue::Text("Boo!".to_string()))])))]
+#[case::non_default_operation(create_operation(SchemaId::new(DEFAULT_SCHEMA_HASH).unwrap(), operation_fields(vec![("message", OperationValue::Text("Boo!".to_string()))])))]
 fn operation_validation(entry: Entry, #[case] operation: Operation, key_pair: KeyPair) {
     let encoded_operation = OperationEncoded::try_from(&operation).unwrap();
     let signed_encoded_entry = sign_and_encode(&entry, &key_pair).unwrap();