+++ p2panda-rs/src/wasm/operation.rs
@@ -11,6 +11,7 @@ use crate::operation::{
     Operation, OperationEncoded, OperationFields as OperationFieldsNonWasm, OperationValue,
     Relation,
 };
+use crate::schema::SchemaId;
 use crate::wasm::error::jserr;
 use crate::wasm::serde::{deserialize_from_js, serialize_to_js};
 use crate::Validate;
@@ -141,7 +142,7 @@ pub fn encode_create_operation(
     schema_hash: String,
     fields: OperationFields,
 ) -> Result<String, JsValue> {
-    let schema = jserr!(Hash::new(&schema_hash));
+    let schema = jserr!(SchemaId::new(&schema_hash));
     let operation = jserr!(Operation::new_create(schema, fields.0));
     let operation_encoded = jserr!(OperationEncoded::try_from(&operation));
     Ok(operation_encoded.as_str().to_owned())
@@ -154,7 +155,7 @@ pub fn encode_update_operation(
     previous_operations: Array,
     fields: OperationFields,
 ) -> Result<String, JsValue> {
-    let schema = jserr!(Hash::new(&schema_hash));
+    let schema = jserr!(SchemaId::new(&schema_hash));
 
     // Decode JsValue into vector of strings
     let prev_op_strings: Vec<String> = jserr!(previous_operations.into_serde());
@@ -177,7 +178,7 @@ pub fn encode_delete_operation(
     schema_hash: String,
     previous_operations: Array,
 ) -> Result<String, JsValue> {
-    let schema = jserr!(Hash::new(&schema_hash));
+    let schema = jserr!(SchemaId::new(&schema_hash));
 
     // Decode JsValue into vector of strings
     let prev_op_strings: Vec<String> = jserr!(previous_operations.into_serde());