+++ p2panda-rs/src/wasm/tests/entry.rs
@@ -5,7 +5,8 @@ use std::convert::TryFrom;
 use wasm_bindgen_test::*;
 
 use crate::hash::Hash;
-use crate::operation::{OperationEncoded, OperationFields, OperationValue};
+use crate::operation::{OperationEncoded, OperationFields, OperationValue, Relation};
+use crate::schema::SchemaId;
 use crate::test_utils::utils::create_operation;
 use crate::wasm::{decode_entry, sign_encode_entry, KeyPair, SignEncodeEntryResult};
 
@@ -14,7 +15,10 @@ wasm_bindgen_test_configure!(run_in_browser);
 #[wasm_bindgen_test]
 fn encodes_decodes_entries() {
     let key_pair = KeyPair::new();
-    let schema = Hash::new_from_bytes(vec![0, 1, 2]).unwrap();
+    let schema = SchemaId::Application(Relation::new(
+        Hash::new_from_bytes(vec![1, 2, 3]).unwrap(),
+        vec![],
+    ));
 
     let mut fields = OperationFields::new();
     fields